Přepis příkazů z aikar commands, oprava bugů a překlepů

+++ src/main/java/cz/craftmania/craftvelocity/api/minetools/objects/MineToolsCache.java
@@ -8,6 +8,6 @@ public class MineToolsCache {
     private @Getter @SerializedName("HIT") boolean hit;
     private @Getter @SerializedName("cache_time") long cacheTime;
     private @Getter @SerializedName("cache_time_left") long cacheTimeLeft;
-    private @Getter @SerializedName("cached_at") long cacheAt;
-    private @Getter @SerializedName("cache_until") long cacheUntil;
+    private @Getter @SerializedName("cached_at") double cacheAt;
+    private @Getter @SerializedName("cache_until") double cacheUntil;
 }
